@@ -17,37 +17,37 @@ dependencies:
   # Core
   intl: any
   collection: any
-  shared_preferences: ^2.1.1
-  uuid: ^3.0.7
-  package_info_plus: ^4.0.1
-  url_launcher: ^6.1.11
+  shared_preferences: ^2.3.2
+  uuid: ^4.4.2
+  package_info_plus: ^8.0.2
+  url_launcher: ^6.3.0
   async: ^2.11.0
-  device_info_plus: ^9.0.1
-  background_json_parser: ^1.0.4 # Json parse işlemini yapar
+  device_info_plus: ^10.1.2
+  background_json_parser: ^1.0.5 # Json parse işlemini yapar
   text_mask: ^1.0.4
 
   # Network
-  http: ^0.13.5
+  http: ^1.2.2
   connectivity: ^3.0.6
-  webview_flutter: ^4.2.1
-  socket_io_client: ^2.0.2
+  webview_flutter: ^4.8.0
+  socket_io_client: ^2.0.3+1
   cached_network_image: ^3.2.3
-  android_id: ^0.3.3
+  android_id: ^0.4.0
 
   # UI
-  cupertino_icons: ^1.0.5
-  flutter_svg: ^2.0.6
-  flutter_local_notifications: ^14.1.0
+  cupertino_icons: ^1.0.8
+  flutter_svg: ^2.0.10+1
+  flutter_local_notifications: ^17.2.2
   banner_image: ^1.0.7
   overlay_kit: ^1.0.8
 
   # Firebase
-  firebase_core: ^2.15.0
-  firebase_messaging: ^14.6.5
-  firebase_dynamic_links: ^5.3.4
+  firebase_core: ^3.3.0
+  firebase_messaging: ^15.0.4
+  firebase_dynamic_links: ^6.0.4
 
   # State manager
-  get: ^4.6.5
+  get: ^4.6.6
 
   # Error handler
   dop_logger:
@@ -56,18 +56,11 @@ dependencies:
       ref: prod-0801
 
 dev_dependencies:
-  # test_coverage: ^0.5.0
-  flutter_test:
-    sdk: flutter
-  integration_test:
-    sdk: flutter
   flutter_lints: any
-  dart_code_metrics: ^5.5.1
   build_runner:
   flutter_gen_runner:
 
 dependency_overrides:
-  async: ^2.8.2
 
 flutter_gen:
   output: lib/app/constants/assets
