@@ -1,239 +1,169 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_base_project/app/extensions/widgets_scale_extension.dart';
-import 'package:flutter_base_project/app/navigation/route/route_factory.dart';
+import '../../navigation/route/route_factory.dart';
 
 /// Uygulama genelinde tüm text style ler buradan çekilmektedir
 /// İsimler 'Size - FontWight - Color' Formatında ayarlanmıştır
-TextStyle get s8W300Dark => Theme.of(MyRouteFactory.context).textTheme.subtitle2!.copyWith(
-      fontWeight: FontWeight.w300,
-      fontSize: 8.horizontalScale,
-    );
+///
+TextStyle get s8w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.displayLarge!.copyWith(fontWeight: FontWeight.w300);
 
-TextStyle get s8W400Dark => Theme.of(MyRouteFactory.context).textTheme.subtitle2!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 8.horizontalScale,
-    );
+TextStyle get s7w400Dark => Theme.of(MyRouteFactory.context).textTheme.displayLarge!.copyWith(fontSize: 7);
 
-TextStyle get s8W500Dark => Theme.of(MyRouteFactory.context).textTheme.subtitle2!.copyWith(
-      fontWeight: FontWeight.w500,
-      fontSize: 8.horizontalScale,
-    );
+TextStyle get s8w400Dark => Theme.of(MyRouteFactory.context).textTheme.displayLarge!;
 
-TextStyle get s8W700Dark => Theme.of(MyRouteFactory.context).textTheme.subtitle2!.copyWith(
-      fontWeight: FontWeight.w700,
-      fontSize: 8.horizontalScale,
-    );
+TextStyle get s8w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.displayLarge!.copyWith(fontWeight: FontWeight.w500);
 
-TextStyle get s10W300Dark => Theme.of(MyRouteFactory.context).textTheme.bodyText2!.copyWith(
-      fontWeight: FontWeight.w300,
-      fontSize: 10.horizontalScale,
-    );
+TextStyle get s8w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.displayLarge!.copyWith(fontWeight: FontWeight.w700);
 
-TextStyle get s10W400Dark => Theme.of(MyRouteFactory.context).textTheme.bodyText2!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 10.horizontalScale,
-    );
+TextStyle get s10w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.displayMedium!.copyWith(fontWeight: FontWeight.w300);
 
-TextStyle get s10W500Dark => Theme.of(MyRouteFactory.context).textTheme.bodyText2!.copyWith(
-      fontWeight: FontWeight.w500,
-      fontSize: 10.horizontalScale,
-    );
+TextStyle get s10w400Dark => Theme.of(MyRouteFactory.context).textTheme.displayMedium!;
 
-TextStyle get s10W700Dark => Theme.of(MyRouteFactory.context).textTheme.bodyText2!.copyWith(
-      fontWeight: FontWeight.w700,
-      fontSize: 10.horizontalScale,
-    );
+TextStyle get s10w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.displayMedium!.copyWith(fontWeight: FontWeight.w500);
 
-TextStyle get s11W300Dark => Theme.of(MyRouteFactory.context).textTheme.caption!.copyWith(
-      fontWeight: FontWeight.w300,
-      fontSize: 11.horizontalScale,
-    );
+TextStyle get s10w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.displayMedium!.copyWith(fontWeight: FontWeight.w700);
 
-TextStyle get s11W400Dark => Theme.of(MyRouteFactory.context).textTheme.caption!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 11.horizontalScale,
-    );
+TextStyle get s11w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.displaySmall!.copyWith(fontWeight: FontWeight.w300);
 
-TextStyle get s11W500Dark => Theme.of(MyRouteFactory.context).textTheme.caption!.copyWith(
-      fontWeight: FontWeight.w500,
-      fontSize: 11.horizontalScale,
-    );
+TextStyle get s11w400Dark => Theme.of(MyRouteFactory.context).textTheme.displaySmall!;
 
-TextStyle get s11W700Dark => Theme.of(MyRouteFactory.context).textTheme.caption!.copyWith(
-      fontWeight: FontWeight.w700,
-      fontSize: 11.horizontalScale,
-    );
+TextStyle get s11w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.displaySmall!.copyWith(fontWeight: FontWeight.w500);
 
-TextStyle get s12W300Dark => Theme.of(MyRouteFactory.context).textTheme.headline1!.copyWith(
-      fontWeight: FontWeight.w300,
-      fontSize: 12.horizontalScale,
-    );
+TextStyle get s11w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.displaySmall!.copyWith(fontWeight: FontWeight.w700);
 
-TextStyle get s12W400Dark => Theme.of(MyRouteFactory.context).textTheme.headline1!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 12.horizontalScale,
-    );
+TextStyle get s12w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.headlineMedium!.copyWith(fontWeight: FontWeight.w300);
 
-TextStyle get s12W500Dark => Theme.of(MyRouteFactory.context).textTheme.headline1!.copyWith(
-      fontWeight: FontWeight.w500,
-      fontSize: 12.horizontalScale,
-    );
+TextStyle get s12w400Dark => Theme.of(MyRouteFactory.context).textTheme.headlineMedium!;
 
-TextStyle get s12W700Dark => Theme.of(MyRouteFactory.context).textTheme.headline1!.copyWith(
-      fontWeight: FontWeight.w700,
-      fontSize: 12.horizontalScale,
-    );
+TextStyle get s12w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.headlineMedium!.copyWith(fontWeight: FontWeight.w500);
 
-TextStyle get s13W300Dark => Theme.of(MyRouteFactory.context).textTheme.headline2!.copyWith(
-      fontWeight: FontWeight.w300,
-      fontSize: 13.horizontalScale,
-    );
+TextStyle get s12w600Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.headlineMedium!.copyWith(fontWeight: FontWeight.w600);
 
-TextStyle get s13W400Dark => Theme.of(MyRouteFactory.context).textTheme.headline2!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 13.horizontalScale,
-    );
+TextStyle get s12w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.headlineMedium!.copyWith(fontWeight: FontWeight.w700);
 
-TextStyle get s13W500Dark => Theme.of(MyRouteFactory.context).textTheme.headline2!.copyWith(
-      fontWeight: FontWeight.w500,
-      fontSize: 13.horizontalScale,
-    );
+TextStyle get s13w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.headlineSmall!.copyWith(fontWeight: FontWeight.w300);
 
-TextStyle get s13W700Dark => Theme.of(MyRouteFactory.context).textTheme.headline2!.copyWith(
-      fontWeight: FontWeight.w700,
-      fontSize: 13.horizontalScale,
-    );
+TextStyle get s13w400Dark => Theme.of(MyRouteFactory.context).textTheme.headlineSmall!;
 
-TextStyle get s14W300Dark => Theme.of(MyRouteFactory.context).textTheme.subtitle1!.copyWith(
-      fontWeight: FontWeight.w300,
-      fontSize: 14.horizontalScale,
-    );
+TextStyle get s13w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.headlineSmall!.copyWith(fontWeight: FontWeight.w500);
 
-TextStyle get s14W400Dark => Theme.of(MyRouteFactory.context).textTheme.subtitle1!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 14.horizontalScale,
-    );
+TextStyle get s13w600Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.headlineSmall!.copyWith(fontWeight: FontWeight.w600);
 
-TextStyle get s14W500Dark => Theme.of(MyRouteFactory.context).textTheme.subtitle1!.copyWith(
-      fontSize: 14.horizontalScale,
-    );
+TextStyle get s13w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.headlineSmall!.copyWith(fontWeight: FontWeight.w700);
 
-TextStyle get s14W600Dark => Theme.of(MyRouteFactory.context).textTheme.subtitle1!.copyWith(
-      fontSize: 14.horizontalScale,
-      fontWeight: FontWeight.w600,
-    );
+TextStyle get s14w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.bodyLarge!.copyWith(fontWeight: FontWeight.w300);
 
-TextStyle get s14W700Dark => Theme.of(MyRouteFactory.context).textTheme.subtitle1!.copyWith(
-      fontWeight: FontWeight.w700,
-      fontSize: 14.horizontalScale,
-    );
+TextStyle get s14w400Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.bodyLarge!.copyWith(fontWeight: FontWeight.w400);
 
-TextStyle get s16W300Dark => Theme.of(MyRouteFactory.context).textTheme.bodyText1!.copyWith(
-      fontWeight: FontWeight.w300,
-      fontSize: 16.horizontalScale,
-    );
+TextStyle get s14w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.bodyLarge!.copyWith(fontWeight: FontWeight.w500);
 
-TextStyle get s16W400Dark => Theme.of(MyRouteFactory.context).textTheme.bodyText1!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 16.horizontalScale,
-    );
+TextStyle get s14w600Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.bodyLarge!.copyWith(fontWeight: FontWeight.w600);
 
-TextStyle get s16W500Dark => Theme.of(MyRouteFactory.context).textTheme.bodyText1!.copyWith(
-      fontWeight: FontWeight.w500,
-      fontSize: 16.horizontalScale,
-    );
+TextStyle get s14w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.bodyLarge!.copyWith(fontWeight: FontWeight.w700);
 
-TextStyle get s16W700Dark => Theme.of(MyRouteFactory.context).textTheme.bodyText1!.copyWith(
-      fontWeight: FontWeight.w700,
-      fontSize: 16.horizontalScale,
-    );
+TextStyle get s22w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.titleLarge!.copyWith(fontWeight: FontWeight.w300);
 
-TextStyle get s15W300Dark => Theme.of(MyRouteFactory.context).textTheme.headline4!.copyWith(
-      fontWeight: FontWeight.w300,
-      fontSize: 15.horizontalScale,
-    );
+TextStyle get s22w400Dark => Theme.of(MyRouteFactory.context).textTheme.titleLarge!;
 
-TextStyle get s15W400Dark => Theme.of(MyRouteFactory.context).textTheme.headline4!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 15.horizontalScale,
-    );
+TextStyle get s22w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.titleLarge!.copyWith(fontWeight: FontWeight.w500);
 
-TextStyle get s15W500Dark => Theme.of(MyRouteFactory.context).textTheme.headline4!.copyWith(
-      fontWeight: FontWeight.w500,
-      fontSize: 15.horizontalScale,
-    );
+TextStyle get s22w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.titleLarge!.copyWith(fontWeight: FontWeight.w700);
 
-TextStyle get s15W700Dark => Theme.of(MyRouteFactory.context).textTheme.headline4!.copyWith(
-      fontWeight: FontWeight.w700,
-      fontSize: 15.horizontalScale,
-    );
+TextStyle get s15w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.titleMedium!.copyWith(fontWeight: FontWeight.w300);
 
-TextStyle get s18W300Dark => Theme.of(MyRouteFactory.context).textTheme.headline3!.copyWith(
-      fontWeight: FontWeight.w300,
-      fontSize: 18.horizontalScale,
-    );
+TextStyle get s15w400Dark => Theme.of(MyRouteFactory.context).textTheme.titleMedium!;
 
-TextStyle get s18W400Dark => Theme.of(MyRouteFactory.context).textTheme.headline3!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 18.horizontalScale,
-    );
+TextStyle get s15w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.titleMedium!.copyWith(fontWeight: FontWeight.w500);
 
-TextStyle get s18W500Dark => Theme.of(MyRouteFactory.context).textTheme.headline3!.copyWith(
-      fontWeight: FontWeight.w500,
-      fontSize: 18.horizontalScale,
-    );
+TextStyle get s15w600Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.titleMedium!.copyWith(fontWeight: FontWeight.w600);
 
-TextStyle get s18W700Dark => Theme.of(MyRouteFactory.context).textTheme.headline3!.copyWith(
-      fontWeight: FontWeight.w700,
-      fontSize: 18.horizontalScale,
-    );
+TextStyle get s15w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.titleMedium!.copyWith(fontWeight: FontWeight.w700);
 
-TextStyle get s20W300Dark => Theme.of(MyRouteFactory.context).textTheme.headline5!.copyWith(
-      fontWeight: FontWeight.w300,
-      fontSize: 20.horizontalScale,
-    );
+TextStyle get s18w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.bodyMedium!.copyWith(fontWeight: FontWeight.w300);
 
-TextStyle get s20W400Dark => Theme.of(MyRouteFactory.context).textTheme.headline5!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 20.horizontalScale,
-    );
+TextStyle get s18w400Dark => Theme.of(MyRouteFactory.context).textTheme.bodyMedium!;
 
-TextStyle get s20W500Dark => Theme.of(MyRouteFactory.context).textTheme.headline5!.copyWith(
-      fontWeight: FontWeight.w500,
-      fontSize: 20.horizontalScale,
-    );
+TextStyle get s18w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.bodyMedium!.copyWith(fontWeight: FontWeight.w500);
 
-TextStyle get s20W700Dark => Theme.of(MyRouteFactory.context).textTheme.headline5!.copyWith(
-      fontWeight: FontWeight.w700,
-      fontSize: 20.horizontalScale,
-    );
+TextStyle get s18w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.bodyMedium!.copyWith(fontWeight: FontWeight.w700);
 
-TextStyle get s22W400Dark => Theme.of(MyRouteFactory.context).textTheme.headline5!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 22.horizontalScale,
-    );
+TextStyle get s20w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.bodySmall!.copyWith(fontWeight: FontWeight.w300);
 
-TextStyle get s26W300Dark => Theme.of(MyRouteFactory.context).textTheme.headline6!.copyWith(
-      fontWeight: FontWeight.w300,
-      fontSize: 26.horizontalScale,
-    );
+TextStyle get s20w400Dark => Theme.of(MyRouteFactory.context).textTheme.bodySmall!;
 
-TextStyle get s26W400Dark => Theme.of(MyRouteFactory.context).textTheme.headline6!.copyWith(
-      fontWeight: FontWeight.w400,
-      fontSize: 26.horizontalScale,
-    );
+TextStyle get s20w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.bodySmall!.copyWith(fontWeight: FontWeight.w500);
 
-TextStyle get s26W500Dark => Theme.of(MyRouteFactory.context).textTheme.headline6!.copyWith(
-      fontWeight: FontWeight.w500,
-      fontSize: 26.horizontalScale,
-    );
+TextStyle get s20w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.bodySmall!.copyWith(fontWeight: FontWeight.w700);
 
-TextStyle get s26W600Dark => Theme.of(MyRouteFactory.context).textTheme.headline6!.copyWith(
-      fontWeight: FontWeight.w600,
-      fontSize: 26.horizontalScale,
-    );
+TextStyle get s16w400Dark => Theme.of(MyRouteFactory.context).textTheme.titleSmall!;
 
-TextStyle get s26W700Dark => Theme.of(MyRouteFactory.context).textTheme.headline6!.copyWith(
-      fontWeight: FontWeight.w700,
-      fontSize: 26.horizontalScale,
-    );
+TextStyle get s16w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.titleSmall!.copyWith(fontWeight: FontWeight.w500);
+
+TextStyle get s16w600Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.titleSmall!.copyWith(fontWeight: FontWeight.w600);
+
+TextStyle get s16w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.titleSmall!.copyWith(fontWeight: FontWeight.w700);
+
+TextStyle get s16w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.titleSmall!.copyWith(fontWeight: FontWeight.w300);
+
+TextStyle get s26w300Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.labelLarge!.copyWith(fontWeight: FontWeight.w300);
+
+TextStyle get s26w400Dark => Theme.of(MyRouteFactory.context).textTheme.labelLarge!;
+
+TextStyle get s26w500Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.labelLarge!.copyWith(fontWeight: FontWeight.w500);
+
+TextStyle get s26w700Dark =>
+    Theme.of(MyRouteFactory.context).textTheme.labelLarge!.copyWith(fontWeight: FontWeight.w700);
+
+TextStyle get s32w300Dark => Theme.of(MyRouteFactory.context).textTheme.labelSmall!.copyWith(
+  fontWeight: FontWeight.w300,
+);
+
+TextStyle get s32w400Dark => Theme.of(MyRouteFactory.context).textTheme.labelSmall!;
+
+TextStyle get s32w500Dark => Theme.of(MyRouteFactory.context).textTheme.labelSmall!.copyWith(
+  fontWeight: FontWeight.w500,
+);
+
+TextStyle get s32w700Dark => Theme.of(MyRouteFactory.context).textTheme.labelSmall!.copyWith(
+  fontWeight: FontWeight.w700,
+);
