@@ -44,7 +44,7 @@ class App extends StatelessWidget {
               ],
               builder: (BuildContext context, Widget? child) {
                 return MediaQuery(
-                  data: MediaQuery.of(context).copyWith(textScaleFactor: 1),
+                  data: MediaQuery.of(context).copyWith(textScaler: const TextScaler.linear(1)),
                   child: child!,
                 );
               },
