@@ -1,16 +1,12 @@
 import 'package:flutter/material.dart';
 
 abstract class AppColors {
-  static Color primary = const Color(0xff344c98);
-  static Color primaryContainer = const Color(0xff415AA8);
-  static Color secondary = const Color(0xffFFCB05);
-  static Color appBarColor = const Color(0xff344c98);
-  static Color onAppBarColor = Colors.white;
-  static Color onSecondary = AppColors.grey.shade800;
-  static Color tertiary = const Color(0xFFEFF8FF);
-  static Color azureishWhite = const Color(0xffE0E4F1);
-  static Color cardColor = Colors.white;
-  static Color lightWhite = const Color(0xffefefef).withOpacity(0.8);
+  AppColors._();
+
+  static const Color primary = Color(0xff344c98);
+  static const Color primaryContainer = Color(0xff415AA8);
+  static const Color secondary = Color(0xFFFFC629);
+  static const Color cardColor = Colors.white;
 
   static MaterialColor grey = const MaterialColor(
     0xFF9E9E9E,
