@@ -11,6 +11,8 @@ abstract class AppTheme {
 
   Color get shadowColor;
 
+  Color get disabledColor;
+
   IconThemeData get accentIconTheme;
 
   IconThemeData get primaryIconTheme;
@@ -51,6 +53,12 @@ abstract class AppTheme {
 
   TabBarTheme get tabBarTheme;
 
+  PopupMenuThemeData get popupMenuThemeData;
+
+  TimePickerThemeData get timePickerThemeData;
+
+  DatePickerThemeData get datePickerTheme;
+
   Color get indicatorColor;
 
   BottomAppBarTheme get bottomAppBarTheme;
@@ -67,35 +75,37 @@ abstract class AppTheme {
 }
 
 ThemeData getTheme<T extends AppTheme>(T theme) => ThemeData(
-      progressIndicatorTheme: theme.progressIndicatorTheme,
-      primaryColorDark: theme.primaryColorDark,
-      primaryColorLight: theme.primaryColorLight,
-      colorScheme: theme.colorScheme,
-      backgroundColor: theme.backgroundColor,
-      scaffoldBackgroundColor: theme.scaffoldBackgroundColor,
-      cardColor: theme.cardColor,
-      shadowColor: theme.shadowColor,
-      cardTheme: theme.cardTheme,
-      fontFamily: fontFamily,
-      textTheme: theme.textTheme,
-      appBarTheme: theme.appBarTheme,
-      bottomAppBarColor: theme.bottomAppBarColor,
-      iconTheme: theme.iconTheme,
-      brightness: theme.brightness,
-      elevatedButtonTheme: theme.elevatedButtonTheme,
-      textSelectionTheme: theme.textSelectionThemeData,
-      inputDecorationTheme: theme.inputDecorationTheme,
-      primaryIconTheme: theme.primaryIconTheme,
-      textButtonTheme: theme.textButtonThemeData,
-      dividerColor: theme.dividerColor,
-      dividerTheme: theme.dividerTheme,
-      outlinedButtonTheme: theme.outlinedButtonThemeData,
-      checkboxTheme: theme.checkboxThemeData,
-      buttonTheme: ButtonThemeData(colorScheme: theme.buttonColorScheme),
-      radioTheme: theme.radioThemeData,
-      tabBarTheme: theme.tabBarTheme,
-      indicatorColor: theme.indicatorColor,
-      bottomAppBarTheme: theme.bottomAppBarTheme,
-      dialogTheme: theme.dialogTheme,
-      hintColor: theme.hintColor,
-    );
+  progressIndicatorTheme: theme.progressIndicatorTheme,
+  primaryColorDark: theme.primaryColorDark,
+  primaryColorLight: theme.primaryColorLight,
+  disabledColor: theme.disabledColor,
+  colorScheme: theme.colorScheme,
+  scaffoldBackgroundColor: theme.scaffoldBackgroundColor,
+  cardColor: theme.cardColor,
+  shadowColor: theme.shadowColor,
+  cardTheme: theme.cardTheme,
+  fontFamily: fontFamily,
+  textTheme: theme.textTheme,
+  appBarTheme: theme.appBarTheme,
+  iconTheme: theme.iconTheme,
+  brightness: theme.brightness,
+  elevatedButtonTheme: theme.elevatedButtonTheme,
+  textSelectionTheme: theme.textSelectionThemeData,
+  inputDecorationTheme: theme.inputDecorationTheme,
+  primaryIconTheme: theme.primaryIconTheme,
+  textButtonTheme: theme.textButtonThemeData,
+  dividerColor: theme.dividerColor,
+  dividerTheme: theme.dividerTheme,
+  outlinedButtonTheme: theme.outlinedButtonThemeData,
+  checkboxTheme: theme.checkboxThemeData,
+  buttonTheme: ButtonThemeData(colorScheme: theme.buttonColorScheme),
+  radioTheme: theme.radioThemeData,
+  tabBarTheme: theme.tabBarTheme,
+  indicatorColor: theme.indicatorColor,
+  bottomAppBarTheme: theme.bottomAppBarTheme,
+  dialogTheme: theme.dialogTheme,
+  hintColor: theme.hintColor,
+  popupMenuTheme: theme.popupMenuThemeData,
+  timePickerTheme: theme.timePickerThemeData,
+  datePickerTheme: theme.datePickerTheme,
+);
